+++ Craftsman/Builders/Tests/IntegrationTests/AddCommandTestBuilder.cs
@@ -67,6 +67,9 @@ public class {classPath.ClassNameWithoutExt} : TestBase
 
     private static string GetPermissionTest(string commandName, Entity entity, string featureName, string permission)
     {
+        if(string.IsNullOrWhiteSpace(permission))
+            return null;
+        
         var fakeCreationDto = FileNames.FakerName(FileNames.GetDtoName(entity.Name, Dto.Creation));
         var fakeEntityVariableName = $"{entity.Name.LowercaseFirstLetter()}One";