+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Tweak producer example
 * No more repository/uow abstraction
 * No more unit tests for user policy handler
+* No more default permission. Must be explicitly provided
 
 ### Fixed